# InfluencerWebExperience/openapi.yaml
openapi: 3.0.3
info:
  title: InfluencerCRB Web Experience Service API
  version: 1.0.0
  description: Brand authentication API for signup, login, logout, and Google/Facebook social auth.
servers:
  - url: https://localhost:8081
paths:
...
  /api/creators/{id}:
    get:
      summary: Get creator by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update creator
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete creator
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
  /api/creator-workflow-approvals:
    get:
      summary: List creator workflow approvals
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignCreatorId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Approval list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "3f96da5f-0380-4f90-bb80-67d57f84d666"
                      campaignCreatorId: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                      reviewRound: 1
                      decision: "approved"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "3f96da5f-0380-4f90-bb80-67d57f84d666"
                        campaignCreatorId: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                        reviewRound: 1
                        decision: "approved"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create creator workflow approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/creator-workflow-approvals/{id}:
    get:
      summary: Get creator workflow approval by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update creator workflow approval
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete creator workflow approval
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /api/creator-workflow-payments:
    get:
      summary: List creator workflow payments
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignCreatorId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Payment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "8a1b5ff2-4be8-4405-80de-4dbefc4b3777"
                      campaignCreatorId: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                      amount: 950.00
                      status: "pending"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "8a1b5ff2-4be8-4405-80de-4dbefc4b3777"
                        campaignCreatorId: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                        amount: 950.00
                        status: "pending"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create creator workflow payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/creator-workflow-payments/{id}:
    get:
      summary: Get creator workflow payment by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update creator workflow payment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete creator workflow payment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
  /api/creator-workflow-events:
    get:
      summary: List creator workflow events
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignCreatorId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListOrPaginatedGenericJson'
              examples:
                nonPaginated:
                  summary: Response when page and size are omitted
                  value:
                    - id: "1e5c29c8-4119-4d28-baa4-c3a1fdc4e888"
                      campaignCreatorId: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                      eventType: "message_sent"
                      actor: "brand_owner"
                paginated:
                  summary: Response when page and/or size is provided
                  value:
                    items:
                      - id: "1e5c29c8-4119-4d28-baa4-c3a1fdc4e888"
                        campaignCreatorId: "a89d2f61-37a3-4b48-95f8-06da45a0a333"
                        eventType: "message_sent"
                        actor: "brand_owner"
                    page: 1
                    size: 20
                    total: 1
                    totalPages: 1
                    hasPrevious: false
                    hasNext: false
    post:
      summary: Create creator workflow event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Created event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
  /api/creator-workflow-events/{id}:
    get:
      summary: Get creator workflow event by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    put:
      summary: Update creator workflow event
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericJson'
      responses:
        '200':
          description: Updated event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericJson'
    delete:
      summary: Delete creator workflow event
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      required: false
      description: 1-based page number. When omitted with size, defaults to 1.
      schema:
        type: integer
        minimum: 1
    SizeQuery:
      name: size
      in: query
      required: false
      description: Page size. When omitted with page, defaults to 20.
      schema:
        type: integer
        minimum: 1
...
  schemas:
    GenericJson:
      type: object
      additionalProperties: true
    PaginatedGenericJson:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GenericJson'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasPrevious:
          type: boolean
        hasNext:
          type: boolean
      required:
        - items
        - page
        - size
        - total
        - totalPages
        - hasPrevious
        - hasNext
    ListOrPaginatedGenericJson:
      oneOf:
        - type: array
          items:
            $ref: '#/components/schemas/GenericJson'
        - $ref: '#/components/schemas/PaginatedGenericJson'
...